elasticsearch/requirements.txt
elasticsearch~=8.10.0
aiohttp~=3.8.0
transformers~=4.33.0
sentence-transformers~=2.2.0
pydantic~=2.4.0
pydantic-settings~=2.0.0
python-dotenv>=1.0.0
srsly>=2.4.6
polars~=0.19.0
codetiming~=1.4.0
rich~=13.6.0
fastapi~=0.104.0
uvicorn>=0.23.0, <1.0.0